[ch-#162584167] moved the server file into server folder
