+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "cross-env NODE_ENV=test && nyc --reporter=html --reporter=text mocha --require babel-register --timeout 10000 ./test/ --exit",
     "start": "babel-node ./server.js",
-    "start:dev": "nodemon --watch server --exec babel-node ./server.js",
+    "start:dev": "nodemon --watch server --exec babel-node ./server/server.js",
     "test:dev": "nyc --reporter=html --reporter=text mocha --compilers js:babel-register ./test*.js --timeout 20000 --exit",
     "coverage": "nyc --reporter=html --reporter=text --reporter=lcov mocha --reporter spec --require babel-register ./test/*.js --exit",
     "coveralls": "npm run coverage -- --report lcovonly && cat coverage/lcov.info | coveralls"
